refactor influence_fn to separate estimator from syntax
