+++ chirho/robust/handlers/estimators.py
@@ -1,13 +1,95 @@
-from typing import Any, Callable, TypeVar
+import warnings
+from typing import Any, Callable
 
+import pyro
 import torch
-from typing_extensions import ParamSpec
 
-from chirho.robust.ops import Functional, Point, influence_fn
+from chirho.robust.ops import Functional, P, Point, S, T, influence_fn
 
-P = ParamSpec("P")
-S = TypeVar("S")
-T = TypeVar("T")
+
+class MonteCarloInfluenceEstimator(pyro.poutine.messenger.Messenger):
+    """
+    Effect handler for approximating efficient influence functions with nested monte carlo.
+    See the MC-EIF estimator in https://arxiv.org/pdf/2403.00158.pdf for details and
+    :func:`~chirho.robust.ops.influence_fn` for example usage.
+
+    .. note::
+
+        * ``functional`` must compose with ``torch.func.jvp``
+        * Since the efficient influence function is approximated using Monte Carlo, the result
+          of this function is stochastic, i.e., evaluating this function on the same ``points``
+          can result in different values. To reduce variance, increase ``num_samples_outer`` and
+          ``num_samples_inner`` in ``linearize_kwargs``.
+        * Currently, ``model`` cannot contain any ``pyro.param`` statements.
+          This issue will be addressed in a future release:
+          https://github.com/BasisResearch/chirho/issues/393.
+        * There are memory leaks when calling this function multiple times due to ``torch.func``.
+          See issue:
+          https://github.com/BasisResearch/chirho/issues/516.
+          To avoid this issue, use ``torch.no_grad()`` as shown in the example above.
+
+    """
+
+    def __init__(self, **linearize_kwargs):
+        self.linearize_kwargs = linearize_kwargs
+        super().__init__()
+
+    def _pyro_influence_fn(self, msg) -> None:
+        functional: Functional[P, S] = msg["args"][0]
+        points: Point[T] = msg["args"][1:]
+
+        if len(points) != 1:
+            raise NotImplementedError(
+                "MonteCarloInfluenceEstimator currently only supports unary functionals"
+            )
+
+        from chirho.robust.internals.linearize import linearize
+        from chirho.robust.internals.utils import make_functional_call
+
+        def _influence_functional(*models: Callable[P, Any]) -> Callable[P, S]:
+            """
+            Functional representing the efficient influence function of ``functional`` at ``points`` .
+
+            :param models: Python callables containing Pyro primitives.
+            :return: efficient influence function for ``functional`` evaluated at ``model`` and ``points``
+            """
+            if len(models) != len(points):
+                raise ValueError("mismatch between number of models and points")
+
+            linearized = linearize(*models, **self.linearize_kwargs)
+            target = functional(*models)
+
+            # TODO check that target_params == model_params
+            assert isinstance(target, torch.nn.Module)
+            target_params, func_target = make_functional_call(target)
+
+            def _fn(*args: P.args, **kwargs: P.kwargs) -> S:
+                """
+                Evaluates the efficient influence function for ``functional`` at each
+                point in ``points``.
+
+                :return: efficient influence function evaluated at each point in ``points`` or averaged
+                """
+                if torch.is_grad_enabled():
+                    warnings.warn(
+                        "Calling influence_fn with torch.grad enabled can lead to memory leaks. "
+                        "Please use torch.no_grad() to avoid this issue. See example in the docstring."
+                    )
+                param_eif = linearized(*points, *args, **kwargs)
+                return torch.vmap(
+                    lambda d: torch.func.jvp(
+                        lambda p: func_target(p, *args, **kwargs),
+                        (target_params,),
+                        (d,),
+                    )[1],
+                    in_dims=0,
+                    randomness="different",
+                )(param_eif)
+
+            return _fn
+
+        msg["value"] = _influence_functional
+        msg["done"] = True
 
 
 def one_step_corrected_estimator(
